% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBN4R.R
\docType{package}
\name{NBN4R}
\alias{NBN4R}
\alias{NBN4R-package}
\title{\pkg{NBN4R}}
\description{
This project enables the R community to access data and tools hosted by the 
NBN Atlas. The goal of the project is to enable basic species and related
information to be queried and output in standard formats for R. 
NBN4R is based around the extensive web services provided 
by the Atlas; see the API link below.
}
\references{
\url{https://api.nbnatlas.org/}
}
